+++ chilitags/cppbuild.sh
@@ -45,6 +45,11 @@ case $PLATFORM in
         make -j4
         make install
         ;;
+    linux-ppc64le)
+        CXX="g++ -m64 -fPIC" $CMAKE -DCMAKE_BUILD_TYPE=Release -DCMAKE_INSTALL_PREFIX=.. -DOpenCV_DIR=$OPENCV_PATH/share/OpenCV/
+        make -j4
+        make install
+        ;;
     macosx-*)
         CXX="clang++ -fPIC" $CMAKE -DCMAKE_BUILD_TYPE=Release -DCMAKE_INSTALL_PREFIX=.. -DOpenCV_DIR=$OPENCV_PATH/share/OpenCV/
         make -j4