+++ libdc1394/cppbuild.sh
@@ -32,6 +32,11 @@ case $PLATFORM in
         make -j4
         make install-strip
         ;;
+    linux-ppc64le)
+        CC="$OLDCC -m64" ./configure --prefix=$INSTALL_PATH
+        make -j4
+        make install-strip
+        ;;
     macosx-*)
         patch -Np1 < ../../../libdc1394-$LIBDC1394_VERSION-macosx.patch
         LIBUSB_CFLAGS=-I/usr/local/include/libusb-1.0/ LIBUSB_LIBS="-L/usr/local/lib/ -lusb-1.0" ./configure --prefix=$INSTALL_PATH