+++ llvm/src/main/java/org/bytedeco/javacpp/presets/clang.java
@@ -25,7 +25,7 @@ package org.bytedeco.javacpp.presets;
 import org.bytedeco.javacpp.annotation.*;
 import org.bytedeco.javacpp.tools.*;
 
-@Properties(inherit = LLVM.class, target = "org.bytedeco.javacpp.clang", value = @Platform(not = {"android", "windows"},
+@Properties(inherit = LLVM.class, target = "org.bytedeco.javacpp.clang", value = @Platform(value = {"linux-x86", "macosx"},
     include = {"<clang-c/Platform.h>", "<clang-c/CXErrorCode.h>", "<clang-c/CXString.h>", "<clang-c/CXCompilationDatabase.h>",
                "<clang-c/BuildSystem.h>", "<clang-c/Index.h>", "<clang-c/Documentation.h>"},
     link = "clang"))